pyproject.toml
[project]
name = "reef-log"
version = "0.1.0"
description = "Local reef tank log with MCP server for Claude — water tests, maintenance events, photo backfill"
authors = [{ name = "findingsimple" }]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "mcp>=1.0.0",
    "Pillow>=10.0.0",
    "click>=8.1.0",
]

[project.scripts]
reef-log = "reef_log.cli:main"
reef-log-mcp = "reef_log.mcp_server:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["reef_log"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
